added test for create goal and floating button components
