+++ src/components/FloatingButton.tsx
@@ -1,41 +1,14 @@
 import { FloatingAction } from "react-native-floating-action";
-import {View, Text, StyleSheet} from 'react-native';
+import {View, StyleSheet} from 'react-native';
 import React from 'react';
-  
-
-const actions = [
-    {
-      text: "Assigned Goal",
-      icon: require('../../assets/assign-user-solid.237x256.png'),
-      buttonSize:50,
-      name: "bt_assignedGoal",
-      position: 1,
-      color : '#492ed1' 
-    },
-    {
-      text: "New Goal",
-      icon: require("../../assets/target.256x256.png"),
-      buttonSize:50,
-      name: "bt_newGoal",
-      position: 2,
-      color : '#492ed1' 
-    },
-    {
-      text: "To Do",
-      icon: require("../../assets/task.201x256.png"),
-      buttonSize:50,
-      name: "bt_todo",
-      position: 3,
-      color : '#492ed1' 
-    },
-  ];
+import FloatingActions from './FloatingActions';
   
 
   export default function FloatingButton() {
     return (
-        <View style={styles.float}>
+        <View testID = "floatingButton" style={styles.float}>
         <FloatingAction
-        actions={actions}
+        actions={FloatingActions}
          color = '#492ed1' 
          position = 'right'
         />
